fix(*): update create course form
